moved raygun client wrapper into services + tied it + set version added
