+++ Resources/config/services.xml
@@ -5,9 +5,10 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="raygun.client.class">Nietonfir\RaygunBundle\Monolog\Handler\Client</parameter>
+        <parameter key="raygun.client.class">Nietonfir\RaygunBundle\Services\Client</parameter>
         <parameter key="nietonfir_raygun.monolog_handler.class">Nietonfir\RaygunBundle\Monolog\Handler\RaygunHandler</parameter>
         <parameter key="nietonfir_raygun.twig_extension.class">Nietonfir\RaygunBundle\Twig\RaygunSetupExtension</parameter>
+        <parameter key="nietonfir_raygun.app_version"></parameter>
 
         <parameter key="nietonfir_raygun.monolog_handler.level" type="constant">Monolog\Logger::ERROR</parameter>
     </parameters>
@@ -27,6 +28,7 @@
             <argument type="service" id="belka.container_support_bundle.raygun_client" />
             <argument>%nietonfir_raygun.monolog_handler.level%</argument>
             <argument>true</argument>
+            <argument>%nietonfir_raygun.app_version%</argument>
         </service>
 
         <service id="nietonfir_raygun.twig_extension" class="%nietonfir_raygun.twig_extension.class%" public="false">
